Add cuRAND support (#75)
This commit adds a cuRAND backend to oneMKL.

The implementation follows closely to that of the existing x86 and
Intel GPU, with the following limitations:

(1) the cuRAND Host API does not support copy-construction for
    generators;
(2) cuRAND does not support initializer lists for seeding
    generators;
(3) cuRAND uses inverse cumulative distribution function (ICDF)
    methods only for quasi-random number generators; and
(4) cuRAND does not support custom ranges on the output sequence
    of random numbers.

Functionality requiring any of 1-3 above are not implemented here
but could be added at a later time. Additional kernels were added to
support (4).