[cublas] Fix build for CUDA 11.x
In CUDA 11 the type of one of the argument of `cublasGemmEx` has been
changed, but CUDA does provide an overload with the previous argument
type for backwards compatibility.

However because it is used as a template argument the introduction of
that overload seemingly broke the template resolution.

This patch simply removes that function from the template arguments and
calls it directly in the intermediary templated function.

This still uses the old signature for `cublasGemmEx` so should work fine
for CUDA 10.x and relies on the compatibility overload for CUDA 11.x.

With this patch oneMKL builds successfully with CUDA 11.4.