Fix out of bounds accesses in imatcopy tests
This patch fixes an issue where the imatcopy tests would occasionally
segfault during the reference computation.

The issue was in the computation of the matrix size, for example taking
the row major case, `lda` is used as the stride for the input matrix, so
in row major, it means that the allocation needs at least `lda` number
of rows. And `lda` is the maximum of `m` and `n`.

So if `lda` is `m`, there's no issue to use `m x n` as the size of the
matrix, but if `lda` is `n`, that also means `m < n`, and therefore a `m
x n` matrix size will lead to out of bound accesses, when using `lda` as
a stride.

So here instead of taking `lda x m` for the row major matrix size, we
need to take `lda x n` to ensure we have enough rows. And similarly, for
row major we need `lda x m`.

This is also consistent with how it's done in
`unit_tests/blas/extensions/imatcopy.cpp`, I'm not sure why these two
tests had it flipped.