remove ENABLE_GENERIC_DEVICE option and add always generic_device to possible devices.
This commit remove the option ENABLE_GENERIC_DEVICE and instead add generic_device
to the backends_table. The check for unsupported_device exception is moved to
table_initializer and to keep it as informative as it is, it is required a
change to the function_tables operator[]. This change allows to use
portBLAS (and in a possible future) all "port" libraries with  any
device supported.