Update vector of device for testing
Previously if some specific backend where enabled the test suite always added a
cpu to the device to run test on, even if another if condition already
added them. This behaviour cause linking time issue if a cpu device is
not available.
This commit removes it and it adds missing pragma to
the other device selection, fixing the linking issue.