Remove dep check
this dep check is overzealous because it enforces that a dependent event cannot be submitted to run on the native device queue but not
completed before a later event it is dependent upon has also been marked
running on the device. This is not part of the sycl spec and
unnecessarily slows down execution.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>